--- src/pages/profile/[userId]/index.tsx
@@ -1,30 +1,0 @@
-import { APIUser } from '@/types/user';
-import TopNavigation from '@/ui/common/TopNavigation';
-import ProfileBookShelf from '@/ui/Profile/ProfileBookshelf';
-import ProfileInfo from '@/ui/Profile/ProfileInfo';
-import { VStack } from '@chakra-ui/react';
-import { GetServerSideProps } from 'next';
-
-const UserProfilePage = ({ userId }: { userId: APIUser['userId'] }) => {
-  return (
-    <VStack justify="center" align="flex-start" w="100%">
-      <TopNavigation pageTitle="" />
-      <VStack justify="flex-start" gap="2rem" w="100%">
-        <ProfileInfo userId={userId} />
-        <ProfileBookShelf userId={userId} />
-      </VStack>
-    </VStack>
-  );
-};
-
-export default UserProfilePage;
-
-export const getServerSideProps: GetServerSideProps = async context => {
-  const { userId } = context.query;
-
-  return {
-    props: {
-      userId: Number(userId),
-    },
-  };
-};
