+++ src/ui/common/TopNavigation/index.tsx
@@ -1,5 +1,5 @@
 import { Box, Flex, Text } from '@chakra-ui/react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import IconButton from '../IconButton';
 
 type PageTitleType = {
